feat(apu): remove BLEED FAULT light behaviour
Based on real-world aircraft observations. Komp has found that the
APU BLEED FAULT light does not illuminate when the APU
doesn't provide bleed air and the APU BLEED pb is ON.
Therefore, until we know what the correct behaviour for the FAULT
light is, we will remove the currently programmed behaviour.